design: MainScrapButton으로 따로 빼서 사용
#82
